Add files via upload
This is a modified file of the original 001_equal_weight_S&P_500.ipynb file. The updates include using pd.concat in place of the deprecated "append" method for data frame. It also uses an updated S&P 500 constituent list pulled out of Wikipedia, which can be used again and again as long as Wikipedia keeps the page, is updated promptly, and does not change the URL. As for the IEX Cloud token, the use policy has changed from during the time the lecture is made (fictitious public version) to now not having such a fictitious public version. The token used in this modification is good at the time of compiling this notebook, but expiring after a week unless I pay for the service (which I don't have plans to in the near future). To use this notebook for your own purpose, register an account with IEX and buy a plan according to your needs.